fix(attn): prevent float16 NaN on V100 by disabling flash_attention + explicit fp32 softmax
V100 does not have a native Flash Attention kernel. PyTorch SDPA falls
back to the math backend, which computes softmax in float16. When the
attention mask contains -inf values (for cross-variable masking in
TimerMultivariateMask), float16 accumulation of (-inf + finite) can
produce NaN, silently corrupting every forward pass.

Fix:
1. Remove --flash_attention from both DDP scripts so the standard
   attention path is always used on V100.
2. In the non-flash path, compute softmax in float32 explicitly:
   softmax(logits.float()).to(logits.dtype)
   This ensures masked-softmax is numerically stable regardless of
   the AMP context.

+++ layers/SelfAttention_Family.py
@@ -108,7 +108,8 @@ class TimeAttention(nn.Module):
             logits = scale * scores + attn_mask
             if getattr(self, 'desta', None) is not None and extra_ctx is not None and ('stats' in extra_ctx):
                 logits = self.desta(logits, extra_ctx['stats'])
-            A = self.dropout(torch.softmax(logits, dim=-1))
+            # 显式 float32 softmax：V100 float16 下 softmax 对含 -inf 的 mask 有数值问题
+            A = self.dropout(torch.softmax(logits.float(), dim=-1).to(logits.dtype))
             V = torch.einsum("bhls,bshd->blhd", A, values)
 
         if self.output_attention:
